user-service/requirements.txt
PyJWT==2.8.0
Flask==2.3.3
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.1
Flask-JWT-Extended==4.6.0